Refactor DTO log for Drawables
Now each implementation of `Drawable` has its own `fromDto()` method that focuses only on that implementation's
needs. `Drawable.fromDto()` simply delegates to each implementation based on the type case of the parent DTO.
