+++ src/main/java/net/rptools/maptool/model/drawing/Cross.java
@@ -74,6 +74,19 @@ public class Cross extends AbstractDrawing {
     return DrawableDto.newBuilder().setCrossDrawable(dto).build();
   }
 
+  public static Cross fromDto(CrossDrawableDto dto) {
+    var id = GUID.valueOf(dto.getId());
+    var startPoint = dto.getStartPoint();
+    var endPoint = dto.getEndPoint();
+    var drawable =
+        new Cross(id, startPoint.getX(), startPoint.getY(), endPoint.getX(), endPoint.getY());
+    if (dto.hasName()) {
+      drawable.setName(dto.getName().getValue());
+    }
+    drawable.setLayer(Zone.Layer.valueOf(dto.getLayer()));
+    return drawable;
+  }
+
   @Override
   public java.awt.Rectangle getBounds(Zone zone) {
     if (bounds == null) {
