+++ src/main/java/net/rptools/maptool/model/drawing/LineCellTemplate.java
@@ -470,4 +470,19 @@ public class LineCellTemplate extends AbstractTemplate {
 
     return DrawableDto.newBuilder().setLineCellTemplate(dto).build();
   }
+
+  public static LineCellTemplate fromDto(LineCellTemplateDto dto) {
+    var id = GUID.valueOf(dto.getId());
+    var drawable = new LineCellTemplate(id);
+    drawable.setRadius(dto.getRadius());
+    var vertex = dto.getVertex();
+    drawable.setVertex(new ZonePoint(vertex.getX(), vertex.getY()));
+    var pathVertex = dto.getPathVertex();
+    drawable.setPathVertex(new ZonePoint(pathVertex.getX(), pathVertex.getY()));
+    if (dto.hasName()) {
+      drawable.setName(dto.getName().getValue());
+    }
+    drawable.setLayer(Zone.Layer.valueOf(dto.getLayer()));
+    return drawable;
+  }
 }
