+++ src/main/java/net/rptools/maptool/model/drawing/LineTemplate.java
@@ -520,4 +520,24 @@ public class LineTemplate extends AbstractTemplate {
 
     return DrawableDto.newBuilder().setLineTemplate(dto).build();
   }
+
+  public static LineTemplate fromDto(LineTemplateDto dto) {
+    var id = GUID.valueOf(dto.getId());
+    var drawable = new LineTemplate(id);
+    drawable.setRadius(dto.getRadius());
+    var vertex = dto.getVertex();
+    drawable.setVertex(new ZonePoint(vertex.getX(), vertex.getY()));
+    if (!dto.getQuadrant().isEmpty()) {
+      drawable.setQuadrant(AbstractTemplate.Quadrant.valueOf(dto.getQuadrant()));
+    }
+    drawable.setMouseSlopeGreater(dto.getMouseSlopeGreater());
+    var pathVertex = dto.getPathVertex();
+    drawable.setPathVertex(new ZonePoint(pathVertex.getX(), pathVertex.getY()));
+    drawable.setDoubleWide(dto.getDoubleWide());
+    if (dto.hasName()) {
+      drawable.setName(dto.getName().getValue());
+    }
+    drawable.setLayer(Zone.Layer.valueOf(dto.getLayer()));
+    return drawable;
+  }
 }
