+++ src/main/java/net/rptools/maptool/model/drawing/RadiusTemplate.java
@@ -175,4 +175,17 @@ public class RadiusTemplate extends AbstractTemplate {
 
     return DrawableDto.newBuilder().setRadiusTemplate(dto).build();
   }
+
+  public static RadiusTemplate fromDto(RadiusTemplateDto dto) {
+    var id = GUID.valueOf(dto.getId());
+    var drawable = new RadiusTemplate(id);
+    drawable.setRadius(dto.getRadius());
+    var vertex = dto.getVertex();
+    drawable.setVertex(new ZonePoint(vertex.getX(), vertex.getY()));
+    if (dto.hasName()) {
+      drawable.setName(dto.getName().getValue());
+    }
+    drawable.setLayer(Zone.Layer.valueOf(dto.getLayer()));
+    return drawable;
+  }
 }
