+++ src/main/java/net/rptools/maptool/model/drawing/WallTemplate.java
@@ -15,9 +15,11 @@
 package net.rptools.maptool.model.drawing;
 
 import com.google.protobuf.StringValue;
+import java.util.ArrayList;
 import java.util.List;
 import net.rptools.maptool.model.CellPoint;
 import net.rptools.maptool.model.GUID;
+import net.rptools.maptool.model.Zone;
 import net.rptools.maptool.model.ZonePoint;
 import net.rptools.maptool.server.proto.drawing.DrawableDto;
 import net.rptools.maptool.server.proto.drawing.WallTemplateDto;
@@ -100,5 +102,29 @@ public class WallTemplate extends LineTemplate {
     for (var point : getPath()) dto.addPoints(point.toDto());
 
     return DrawableDto.newBuilder().setWallTemplate(dto).build();
+  }
+
+  public static WallTemplate fromDto(WallTemplateDto dto) {
+    var id = GUID.valueOf(dto.getId());
+    var drawable = new WallTemplate(id);
+    drawable.setRadius(dto.getRadius());
+    var vertex = dto.getVertex();
+    drawable.setVertex(new ZonePoint(vertex.getX(), vertex.getY()));
+    drawable.setMouseSlopeGreater(dto.getMouseSlopeGreater());
+    var pathVertex = dto.getPathVertex();
+    drawable.setPathVertex(new ZonePoint(pathVertex.getX(), pathVertex.getY()));
+    drawable.setDoubleWide(dto.getDoubleWide());
+    if (dto.hasName()) {
+      drawable.setName(dto.getName().getValue());
+    }
+    drawable.setLayer(Zone.Layer.valueOf(dto.getLayer()));
+
+    var cellpoints = new ArrayList<CellPoint>();
+    for (var point : dto.getPointsList()) {
+      cellpoints.add(new CellPoint(point.getX(), point.getY()));
+    }
+    drawable.setPath(cellpoints);
+
+    return drawable;
   }
 }
